Do not include a .patch, but rather the patched MOC source code.
While just including a .patch file looks very clean and minimalistic,
it is a major PITA to actually work with the code due to the need to
constantly re-apply the patch after cleaning the build directory,
and re-generating the patch after each source change.

Instead, apply the changes directly over the MOC source code and
provide a way to generate the patch from there.