Recycle synthesizers where possible.
Creating a synthesizer using new_fluid_synth can be costly, since it
involves loading the SoundFont, and some such as Arachno weight >100MB.

There are two common scenarios where this can happen:
- If the user is browsing a folder, and begins skimming between MIDIs,
  each track change will release then load the SoundFont anew.
- When doing continuous play between two MIDIs, at the end of the first
  MIDI, MOC will tell us to create a 2nd synthesizer for pre-buffering.
  Shortly after, the track will finish and we will release the 1st one.
  This will happen over and over during the playback of a folder.

Note that we can not just have a single global synthesizer, because we
need two for pre-buffering the next track during for continuous play.

However, what we can do, is pool a single 'available' synthesizer on
fluidsynth_close(), and reuse it for a future fluidsynth_open() call.

With this strategy, both the 'skimming various MIDIs on a folder' and
the 'continuous play' scenario do not need to constantly cycle between
new synthesizers, but can just juggle with the single 'free' one.