Eventually reclaim unused FluidSynth synthesizers.
A minor problem with the logic for recycling synthesizer instances, is
that if the user plays a single MIDI file, the pooled synthesizer
(and its potentially quite heavy SoundFont) will stay loaded in memory,
even if the user never intends to play a MIDI again.

Technically, the pooled synthesizer is not leaked (it is cleaned on
plugin shutdown) but it'd be better to reclaim it after some time.

This implements a relatively simple cleanup logic that will free the
pooled synthesizer after 5 minutes of being unused.