Make SoundFonts selectable from the user interface.
Instead of forcing the user to select a SoundFont at startup, allow him to
select a font from the user interface. It will instantly 'play', and cause
the next MIDI files to be played with that soundfont as a side effect.

Note to self: After implementing this, I think it is a bad idea... it
doesn't work well to select a single SoundFont from a folder with many,
introduces hidden state and does not generalize to loading multiple SF's.